refactor(currencies): promote transaction-form to v2/components
フォームコンポーネントの配置を currency-form と対称化。transaction-form は detail
ページ専用だったが、currency-form 同様に feature の components/ 配下へ昇格する。

- pages/currency-detail-page/transaction-form/（type-combobox/ と __tests__ 同梱）を
  v2/components/transaction-form/ へ移設。
- currency-detail-page.tsx の TransactionFormV2 import を相対 → 絶対(components) に更新。
- 移設に伴いテストの絶対 import / vi.mock パスを新パスへ更新。

挙動・DOM 不変。テスト件数も不変（web-dom 16/268, web-node 2/28 = 19/300）。
