+++ apps/web/src/features/currencies/v2/components/transaction-form/__tests__/transaction-form.test.tsx
@@ -15,7 +15,7 @@ vi.mock("@/features/currencies/hooks/use-transaction-types", () => ({
 	useTransactionTypes: hoisted.useTransactionTypes,
 }));
 
-import { TransactionFormV2 } from "@/features/currencies/v2/pages/currency-detail-page/transaction-form";
+import { TransactionFormV2 } from "@/features/currencies/v2/components/transaction-form";
 
 describe("TransactionFormV2", () => {
 	beforeEach(() => {
