+++ apps/web/src/features/currencies/v2/pages/currency-detail-page/__tests__/currency-detail-page.test.tsx
@@ -100,40 +100,37 @@ vi.mock("@/features/currencies/v2/components/currency-form", () => ({
 	),
 }));
 
-vi.mock(
-	"@/features/currencies/v2/pages/currency-detail-page/transaction-form",
-	() => ({
-		TransactionFormV2: ({
-			defaultValues,
-			onSubmit,
-		}: {
-			defaultValues?: { amount?: number };
-			onSubmit: (values: {
-				amount: number;
-				transactedAt: string;
-				transactionTypeId: string;
-			}) => void;
-		}) => (
-			<div
-				data-default-amount={String(defaultValues?.amount ?? "")}
-				data-testid="transaction-form-stub"
+vi.mock("@/features/currencies/v2/components/transaction-form", () => ({
+	TransactionFormV2: ({
+		defaultValues,
+		onSubmit,
+	}: {
+		defaultValues?: { amount?: number };
+		onSubmit: (values: {
+			amount: number;
+			transactedAt: string;
+			transactionTypeId: string;
+		}) => void;
+	}) => (
+		<div
+			data-default-amount={String(defaultValues?.amount ?? "")}
+			data-testid="transaction-form-stub"
+		>
+			<button
+				onClick={() =>
+					onSubmit({
+						amount: 5,
+						transactedAt: "2026-01-01",
+						transactionTypeId: "t1",
+					})
+				}
+				type="button"
 			>
-				<button
-					onClick={() =>
-						onSubmit({
-							amount: 5,
-							transactedAt: "2026-01-01",
-							transactionTypeId: "t1",
-						})
-					}
-					type="button"
-				>
-					stub-submit-transaction
-				</button>
-			</div>
-		),
-	})
-);
+				stub-submit-transaction
+			</button>
+		</div>
+	),
+}));
 
 vi.mock(
 	"@/features/currencies/v2/pages/currency-detail-page/transaction-list",
